agent1/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
langgraph
langchain-google-genai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic>=2.0.0
a2a-sdk[http-server]
